# api-backend/fripay/public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FriPay API Documentation",
        "version": "1.0.0",
        "description": "API FriPay - Transferts inter-op\u00e9rateurs Mobile Money pour le B\u00e9nin. Endpoint de documentation: `\/api\/docs\/api`"
    },
    "servers": [
        {
            "url": "http:\/\/payments.fripay.local\/api"
        }
    ],
    "paths": {
...
        "\/v1\/qr\/transfer": {
            "post": {
                "operationId": "offlineQr.transfer",
                "description": "Transf\u00e8re la propri\u00e9t\u00e9 d'un QR Code \u00e0 un autre utilisateur\n(ex: un marchand). Le destinataire est identifi\u00e9 par son num\u00e9ro\nde t\u00e9l\u00e9phone. Seul le propri\u00e9taire actuel du QR Code peut le transf\u00e9rer.",
                "summary": "Transf\u00e9rer un QR Code \u00e0 un autre utilisateur",
                "tags": [
                    "QR Codes Hors-ligne",
                    "OfflineQr"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    },
                                    "recipient_phone": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "uuid",
                                    "recipient_phone"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "QR Code transf\u00e9r\u00e9"
                                        },
                                        "uuid": {
                                            "type": "string"
                                        },
                                        "amount": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "uuid",
                                        "amount"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "const": "RECIPIENT_NOT_FOUND"
                                                }
                                            },
                                            "required": [
                                                "error"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "const": "QR_NOT_FOUND"
                                                }
                                            },
                                            "required": [
                                                "error"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "const": "NOT_OWNER"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#\/components\/responses\/ValidationException"
                    },
                    "401": {
                        "$ref": "#\/components\/responses\/AuthenticationException"
                    }
                }
            }
        },
...
        "\/v1\/transfers": {
            "post": {
                "operationId": "transfer.initiate",
                "tags": [
                    "Transfer"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/InitiateTransferRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "transaction_id": {
                                            "type": "string"
                                        },
                                        "reference": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "transaction_id",
                                        "reference",
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#\/components\/responses\/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#\/components\/responses\/ValidationException"
                    }
                }
            },
            "get": {
                "operationId": "transfer.index",
                "tags": [
                    "Transfer"
                ],
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "-initiated_at"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "page": {
                                                    "type": "string"
                                                },
                                                "size": {
                                                    "type": "string"
                                                },
                                                "total": {
                                                    "type": "string"
                                                },
                                                "has_next": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "page",
                                                "size",
                                                "total",
                                                "has_next"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#\/components\/responses\/AuthenticationException"
                    }
                }
            }
        },
...
    },
...
}